// src/server/model/stockSummaryReportInput2.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { IAttributeGroupLine } from './iAttributeGroupLine';

export interface StockSummaryReportInput2 { 
    group?: string;
    date?: Date;
    location?: string;
    useMaxQuantity?: boolean;
    maxQuantity?: number;
    useMinQuantity?: boolean;
    minQuantity?: number;
    showZeroQty?: boolean;
    showOpening?: boolean;
    showStockValue?: boolean;
    showManufacture?: boolean;
    showTaxInfo?: boolean;
    showItemGroup?: boolean;
    attrGroupLine?: IAttributeGroupLine;
}